Expand protocol fields to 16-bit and increase packet size
- Widen type, address, status fields from 8-bit to 16-bit in packet structs
  (dsTxPacket, dsRxPacket), increasing packet size from 6/7 to 8/9 bytes
- Add __attribute__((packed)) to ds_reply_values, ds_type_t, ds_sys_command_t,
  ds_control_interface_t, and discovery enum for correct memory layout
- Add __attribute__((aligned(4))) to packet unions for natural alignment of
  the 32-bit value field (atomic access without critical section)
- Update all address/type parameters across dsParameters, dsRegisters, dsCLI,
  datastream, dsTCP, dsUDP from uint32_t to uint16_t
- Move ds_sys_command_t reserved range to 65000-65535 to match uint16_t space
- Add DS_REPLY_VAL_DUMMY and ds_type_DUMMY sentinels to enforce 16-bit enum size
- Update static assertions from 256 to 65536 address space limit

+++ src/datastream.c
@@ -308,7 +308,7 @@ WEAK bool dsProcessUserSysCommand(const dsRxPacket *inputPacket, dsTxPacket *out
     return false;
 }
 
-void dsPacketSizeError(uint32_t inPacketSize, dsTxPacket *outPacket)
+void dsPacketSizeError(uint16_t inPacketSize, dsTxPacket *outPacket)
 {
     // This function is called when the received packet size is not equal to DATASTREAM_INPUT_SIZE
     // It will return an error message in the output packet
